hnl_dk: Updated to the latest openbsd ksh fixes

+++ history.c
@@ -340,7 +340,7 @@ hist_get(const char *str, int approx, int allow_cur)
 
         if (getn(str, &n)) {
                 hp = histptr + (n < 0 ? n : (n - hist_source->line));
-                if (hp < history) {
+                if ((long)hp < (long)history) {
                         if (approx)
                                 hp = hist_get_oldest();
                         else {
